chore: move template parsing to main process also integrated two input region APIs
