+++ gui/src/routes/(transparent)/menu/template.ts
@@ -1,129 +1,16 @@
-import type { BtnImages, BtnState, LayoutNode, ElementTemplate } from "./types";
+import type { ElementTemplate } from "$bridge/contracts/menu-api";
 
 // Template cache: style path → parsed template
-const templateCache = new Map<string, ElementTemplate>();
+const templateCache: Record<string, ElementTemplate> = {};
 
-/** Convert #AABBGGRR (ABGR) to CSS #RRGGBBAA. */
-function abgrToCss(c: string): string {
-  if (c.length === 9 && c[0] === "#") {
-    // input: #AA BB GG RR  (indices 1-2, 3-4, 5-6, 7-8)
-    return `#${c.slice(7)}${c.slice(5, 7)}${c.slice(3, 5)}${c.slice(1, 3)}`;
-  }
-  return c;
-}
-
-/**
- * Parse a DUI attribute string like:
- * normalimage="file='btn/play.svg' svg_color='#ff483228'" hotimage="..."
- * into structured BtnImages.
- */
-export function parseBtnUrl(url: string): BtnImages | null {
-  const stateRe = /(normalimage|hotimage|pushedimage|disabledimage)="([^"]*)"/g;
-  const states: Record<string, BtnState> = {};
-
-  let m;
-  while ((m = stateRe.exec(url)) !== null) {
-    const key = m[1].replace("image", "");
-    const attrs = m[2];
-    const fileMatch = attrs.match(/file='([^']*)'/);
-    if (!fileMatch) continue;
-    const uri = fileMatch[1];
-    const colorMatch = attrs.match(/svg_color='([^']*)'/);
-    const color = colorMatch?.[1];
-    states[key] = { uri, color: color ? abgrToCss(color) : undefined };
-  }
-
-  if (!states.normal) return null;
-  return {
-    normal: states.normal,
-    hot: states.hot,
-    pushed: states.pushed,
-    disabled: states.disabled,
-  };
-}
-
-function parseNum(el: Element, attr: string): number | undefined {
-  const v = el.getAttribute(attr);
-  return v != null ? Number(v) : undefined;
-}
-
-function parseLayoutNode(
-  el: Element,
-  counter: { i: number }
-): LayoutNode | null {
-  const tag = el.tagName;
-  if (tag === "HorizontalLayout") {
-    return { type: "horizontal", children: parseChildren(el, counter) };
-  }
-  if (tag === "VerticalLayout") {
-    return { type: "vertical", children: parseChildren(el, counter) };
-  }
-  if (tag === "Container") {
-    return {
-      type: "container",
-      width: parseNum(el, "width"),
-      height: parseNum(el, "height"),
-      children: parseChildren(el, counter),
-    };
-  }
-  if (tag === "Control") {
-    return {
-      type: "control",
-      width: parseNum(el, "width"),
-      height: parseNum(el, "height"),
-    };
-  }
-  if (tag === "Button") {
-    return {
-      type: "button",
-      width: parseNum(el, "width") ?? 24,
-      height: parseNum(el, "height") ?? 24,
-      index: counter.i++,
-    };
-  }
-  // MenuButton and MenuLabel are part of the default element.xml template — skip them
-  return null;
-}
-
-function parseChildren(parent: Element, counter: { i: number }): LayoutNode[] {
-  const nodes: LayoutNode[] = [];
-  for (const child of parent.children) {
-    const node = parseLayoutNode(child, counter);
-    if (node) nodes.push(node);
-  }
-  return nodes;
-}
-
-function parseElementTemplate(xml: string): ElementTemplate | null {
-  const doc = new DOMParser().parseFromString(xml, "text/xml");
-  const menuEl = doc.querySelector("MenuElement");
-  if (!menuEl) return null;
-
-  const layoutEl = menuEl.querySelector("MenuElementLayout");
-  if (!layoutEl) return null;
-
-  return {
-    height: parseNum(menuEl, "height") ?? 30,
-    minWidth: parseNum(menuEl, "minwidth") ?? 0,
-    maxWidth: parseNum(menuEl, "maxwidth") ?? 300,
-    layout: {
-      type: "horizontal",
-      children: parseChildren(layoutEl, { i: 0 }),
-    },
-  };
-}
-
-/** Parse and cache all raw XML templates sent from the main process. */
-export function loadTemplates(rawTemplates: Record<string, string>) {
-  for (const [style, xml] of Object.entries(rawTemplates)) {
-    if (!templateCache.has(style)) {
-      const tpl = parseElementTemplate(xml);
-      if (tpl) templateCache.set(style, tpl);
-    }
-  }
+/** Store pre-parsed templates from the main process. */
+export function loadTemplates(
+  parsedTemplates: Record<string, ElementTemplate>
+) {
+  Object.assign(templateCache, parsedTemplates);
 }
 
 /** Get cached template synchronously (must be preloaded). */
 export function getCachedTemplate(style: string): ElementTemplate | null {
-  return templateCache.get(style) ?? null;
+  return templateCache[style] ?? null;
 }
