+++ gui/src/routes/(transparent)/menu/types.ts
@@ -1,6 +1,8 @@
+import type { BtnImages } from "$bridge/contracts/menu-api";
+
 export interface MenuItemBtn {
   id: string;
-  url: string;
+  images: BtnImages | null;
   enable: boolean;
 }
 
@@ -19,33 +21,9 @@ export interface MenuItem {
   btns?: MenuItemBtn[];
 }
 
-export interface BtnState {
-  uri: string;
-  color?: string;
-}
-
-export interface BtnImages {
-  normal: BtnState;
-  hot?: BtnState;
-  pushed?: BtnState;
-  disabled?: BtnState;
-}
-
-export type LayoutNode =
-  | { type: "horizontal"; children: LayoutNode[] }
-  | { type: "vertical"; children: LayoutNode[] }
-  | {
-      type: "container";
-      width?: number;
-      height?: number;
-      children: LayoutNode[];
-    }
-  | { type: "control"; width?: number; height?: number }
-  | { type: "button"; width: number; height: number; index: number };
-
-export interface ElementTemplate {
-  height: number;
-  minWidth: number;
-  maxWidth: number;
-  layout: LayoutNode;
-}
+export type {
+  BtnState,
+  BtnImages,
+  ElementTemplate,
+  LayoutNode,
+} from "$bridge/contracts/menu-api";
