+++ package.json
@@ -47,6 +47,7 @@
     "@types/unzipper": "^0.10.11",
     "@typescript-eslint/eslint-plugin": "^8.59.0",
     "@typescript-eslint/parser": "^8.59.0",
+    "@xmldom/xmldom": "^0.9.10",
     "@yxim/nim-web-sdk": "^9.21.12",
     "cookie": "^1.1.1",
     "electron": "41.4.0",
