+++ src/bridge/contracts/input-region-api.ts
@@ -1,0 +1,11 @@
+export interface InputRegion {
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
+
+export interface InputRegionContract {
+  platform: NodeJS.Platform;
+  setInputRegions(regions: InputRegion[]): Promise<void>;
+}
