+++ src/bridge/contracts/menu-api.ts
@@ -1,10 +1,17 @@
 import type { MenuSkin } from "../../main/menu/types";
+import type { ElementTemplate } from "../../main/skin/dui";
 
 export type { MenuSkin } from "../../main/menu/types";
+export type {
+  BtnImages,
+  BtnState,
+  ElementTemplate,
+  LayoutNode,
+} from "../../main/skin/dui";
 
 export interface MenuPullResult {
   items: unknown[];
-  templates: Record<string, string>;
+  templates: Record<string, ElementTemplate>;
   colors: MenuSkin;
   cursorX?: number;
   cursorY?: number;
@@ -18,7 +25,7 @@ export interface MenuContract {
     show(
       callback: (
         items: unknown[],
-        templates: Record<string, string>,
+        templates: Record<string, ElementTemplate>,
         cursorX: number,
         cursorY: number,
         colors: MenuSkin
@@ -34,7 +41,7 @@ export interface MenuContract {
   close(): Promise<void>;
   openSubmenu(
     items: unknown[],
-    templates: Record<string, string>,
+    templates: Record<string, ElementTemplate>,
     x: number,
     y: number
   ): Promise<void>;
