+++ src/windows/desktop-lyrics.ts
@@ -1,5 +1,6 @@
 import { exposeApi } from "../bridge/preload";
 
-exposeApi("desktopLyrics", {
+exposeApi("desktopLyrics");
+exposeApi("inputRegion", {
   platform: process.platform,
 });
