+++ src/windows/mini-player.ts
@@ -1,5 +1,6 @@
 import { exposeApi } from "../bridge/preload";
 
-exposeApi("miniPlayer", {
+exposeApi("miniPlayer");
+exposeApi("inputRegion", {
   platform: process.platform,
 });
